# MixerHMIInterface.swagger.yml
swagger: '2.0'
info:
  version: 0.0.1
  title: payment
paths:
  
  /drinks:
    get:
      description: gets all drinks
      parameters:
        - name: after
          in: query
          description: createdDate
          required: false
          type: string
        - name: ingredients
          in: query
          description: array of possible ingredients
          type: array
          items:
            type: string
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Drink'
    
  '/drinks/{drinkId}':
    get:
      description: gets one recipes
      parameters:
        - name: drinkId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Drink'
    
  '/users/{userId}':
    get:
      parameters:
        - in: path
          name: userId
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
            
  /orders:
    post:
      produces:
          - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        '201':
          description: Order succesfully placed
          headers:
            location:
              description: The Order ID
              type: string
  /order/{id}:
    get:
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
      responses:
       '200':
          description: the requested order
          schema:
            $ref: "#/definitions/Order"
            
    delete:
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
      responses:
       '200':
          description: the requested order
          schema:
            $ref: "#/definitions/Order"
  
  '/orders/{id}/paymentRequest':
    get:
      produces:
        - text/plain
      parameters:
        - in: path
          name: id
          required: true
          type: string
      responses:
        '200':
          description: the payment Request
          
  '/order/{id}/payment':
    put:
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
        - in: body
          name: body
          schema:
            type: string
          required: true
      responses:
       '200':
          description: ok
          
  '/order/{id}/productionStart':
    put:
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
      responses:
       '200':
          description: ok
       '500':
          description: Order is not in correct state
       '404':
          description: Order Id is not listed
          
  '/order/{id}/resume':
    put:
      description:
        Used to resume the production for this order in case of an errror
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
      responses:
       '200':
          description: ok
       '404':
          description: Order Id is not listed
          
          
  '/admin/startConfirm':
    post:
      description:
        Used to confirm the production start for the first order in the order queue
        
      produces:
        - application/json
      responses:
        '200':
          description: ok
  '/admin/production/pause':
    post:
      description:
        Used to pause the production. can only be applied, when the machine is not pumping
        
      produces:
        - application/json
      responses:
        '200':
          description: ok
          
  '/admin/production/resume':
    post:
      description:
        Used to resume the production. can only be applied, when the machine is in pause mode
        
      produces:
        - application/json
      responses:
        '200':
          description: ok
          
  '/components':
    get:
      description:
        gets the available components for the pumps
      produces:
        - application/json
        
      responses:
        '200':
          description: The list of components
          schema:
            type: array
            items:
              $ref: "#/definitions/Component"
  
  '/admin/pumps/service':
    post:
      description:
        Used to activate the service mode for the pumps true on false off
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: string
      produces:
        - application/json
      responses:
        '200':
          description: ok
  
  
  '/admin/pumps':
    get:
      description:
        gets the current configuration for the pumps
      produces:
        - application/json
        
      responses:
        '200':
          description: The list of components
          schema:
            type: array
            items:
              $ref: "#/definitions/Pump"
              
  '/admin/pumps/{nr}':
    put:
      description:
        change the component for the given pump nr
      parameters:
        - in: path
          name: nr
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Component"
      responses:
        '201':
          description: ok
  '/admin/pumps/{nr}/active':
    put:
      description:
        activate or deactivate a pump in service mode by sending true or false
      parameters:
        - in: path
          name: nr
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: string
      responses:
        '201':
          description: ok
          
          
definitions:
  
    
  Component:
    type: object
    properties:
      id: 
        type: string
      name:
        type: string
        
  Pump:
    type: object
    properties:
      nr:
        type: number
      component:
        $ref: "#/definitions/Component"
        
  Price:
    type: object
    properties:
      amount:
        type: number
        format: double
      currency:
        type: string
  
  Rating:
    type: object
    properties:
      value:
        type: number
      min:
        type: number
      max:
        type: number
      count:
        type: number
  User:
    type: object
    properties:
      id:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      thumbnail:
        type: string
        description: The Thumbnail jpg in base64
      imageRef:
        type: string
        description: The reference to the image
  Drink:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      description:
        type: string
      thumbnail:
        type: string
        format: byte
        description: The Thumbnail jpg in base64
      imageRef:
        type: string
        description: The reference to the image
      authorId:
        type: string
      createdAt:
        type: string
        format: date
      updatedAt:
        type: string
        format: date
      rating:
        $ref: '#/definitions/Rating'
      retailPrice:
        $ref: '#/definitions/Price'

  Order:
    type: object
    properties:
      recipeId:
        type: string
      customerName:
        type: string
      
        